* dedicated server fix (rewrite to use packets instead of handlers this will also mean that we can migrate to non MUI stuff later or use commands instead of items and each scan is now unique)
