+++ src/main/java/bioast/mods/gt6scan/network/OreData.java
@@ -1,13 +1,13 @@
 package bioast.mods.gt6scan.network;
 
 public class OreData {
-    public final int x, y, z;
-    public final short matID;
+	public final int x, y, z;
+	public final short matID;
 
-    public OreData(int x, int y, int z, short matID) {
-        this.x = x;
-        this.y = y;
-        this.z = z;
-        this.matID = matID;
-    }
+	public OreData(int x, int y, int z, short matID) {
+		this.x = x;
+		this.y = y;
+		this.z = z;
+		this.matID = matID;
+	}
 }
